class3s: separate out factorial() function
