+++ class3/build.sh
@@ -2,4 +2,5 @@
 
 gcc -Wall -c hello.c
 gcc -Wall -c greeting.c
-gcc hello.o greeting.o -o hello
+gcc -Wall -c factorial.c
+gcc hello.o greeting.o factorial.o -o hello
