+++ class3/factorial.c
@@ -1,0 +1,15 @@
+
+#include "hello.h"
+
+int factorial(int n)
+{
+    int i;
+    int retval = 1;
+
+    for (i = 1; i <= n; i++)
+    {
+        retval *= i;
+    }
+
+    return retval;
+}
