+++ class3/hello.c
@@ -3,19 +3,6 @@
 
 #include <stdio.h>
 
-int factorial(int n)
-{
-    int i;
-    int retval = 1;
-
-    for (i = 1; i <= n; i++)
-    {
-        retval *= i;
-    }
-
-    return retval;
-}
-
 int main(int argc, char **argv)
 {
     greeting("everyone");
