+++ class3/hello.h
@@ -2,5 +2,6 @@
 #define HELLO_H
 
 void greeting(const char *name);
+int factorial(int n);
 
 #endif
